fix(clever): install assertthat library
This is required by the `mailR` library that is required by
`remaputils`.

Before this fix, DECAF Clever runtime build did not manage to install
`mailR` which was not a fatal error for the build process. Therefore,
we did not notice it.

I noticed that when I was going through build logs.